+++ ledger/types/reserved_fund.go
@@ -16,14 +16,14 @@ type ReservedFund struct {
 	InitialFund     Coins            `json:"initial_fund"`
 	UsedFund        Coins            `json:"used_fund"`
 	ResourceIDs     [][]byte         `json:"resource_ids"` // List of resource ID
-	EndBlockHeight  uint32           `json:"end_block_height"`
-	ReserveSequence int              `json:"reserve_sequence"` // sequence number of the corresponding ReserveFundTx transaction
+	EndBlockHeight  uint64           `json:"end_block_height"`
+	ReserveSequence uint64           `json:"reserve_sequence"` // sequence number of the corresponding ReserveFundTx transaction
 	TransferRecords []TransferRecord `json:"transfer_records"` // signed ServerPaymentTransactions
 }
 
 // TODO: this implementation is not very efficient
-func (reservedFund *ReservedFund) VerifyPaymentSequence(targetAddress common.Address, paymentSequence int) error {
-	currentPaymentSequence := 0
+func (reservedFund *ReservedFund) VerifyPaymentSequence(targetAddress common.Address, paymentSequence uint64) error {
+	currentPaymentSequence := uint64(0)
 	for _, transferRecord := range reservedFund.TransferRecords {
 		transferRecordTargetAddr := transferRecord.ServicePayment.Target.Address
 		if targetAddress == transferRecordTargetAddr {
